# CUDA: faster multi GPU synchronization